requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
cryptography==42.0.8
reportlab==4.2.2
python-dotenv==1.0.1
Pillow==10.4.0
